+++ packages/dear-imgui/cpp/app/CMakeLists.txt
@@ -5,7 +5,7 @@ set(CMAKE_TOOLCHAIN_FILE ../deps/vcpkg/scripts/buildsystems/vcpkg.cmake)
 
 set(VCPKG_TARGET_TRIPLET "x64-windows")
 
-project(reactDearImgui LANGUAGES C CXX)
+project(xframes LANGUAGES C CXX)
 
 set(CMAKE_CXX_STANDARD 23)
 
@@ -35,48 +35,48 @@ file(GLOB YOGA_SRC CONFIGURE_DEPENDS
         ${DEPS}/yoga/yoga/*.cpp
         ${DEPS}/yoga/yoga/**/*.cpp)
 
-set(REACT_IMGUI_BASE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/../app)
+set(XFRAMES_BASE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/../app)
 
-set(REACT_IMGUI_SRC
-        ${REACT_IMGUI_BASE_PATH}/src/reactimgui.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/imgui_renderer.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/implot_renderer.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/shared.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/color_helpers.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/imgui_helpers.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/yoga_helpers.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/element/layout_node.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/element/element.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/widget.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/styled_widget.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/button.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/checkbox.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/child.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/clipped_multi_line_text_renderer.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/collapsing_header.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/combo.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/group.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/image.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/input_text.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/item_tooltip.cpp
-#        ${REACT_IMGUI_BASE_PATH}/src/widget/map_view.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/multi_slider.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/plot_line.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/plot_candlestick.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/separator.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/separator_text.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/slider.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/table.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/tabs.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/text.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/text_wrap.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/tree_node.cpp
-        ${REACT_IMGUI_BASE_PATH}/src/widget/window.cpp
+set(XFRAMES_SRC
+        ${XFRAMES_BASE_PATH}/src/reactimgui.cpp
+        ${XFRAMES_BASE_PATH}/src/imgui_renderer.cpp
+        ${XFRAMES_BASE_PATH}/src/implot_renderer.cpp
+        ${XFRAMES_BASE_PATH}/src/shared.cpp
+        ${XFRAMES_BASE_PATH}/src/color_helpers.cpp
+        ${XFRAMES_BASE_PATH}/src/imgui_helpers.cpp
+        ${XFRAMES_BASE_PATH}/src/yoga_helpers.cpp
+        ${XFRAMES_BASE_PATH}/src/element/layout_node.cpp
+        ${XFRAMES_BASE_PATH}/src/element/element.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/widget.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/styled_widget.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/button.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/checkbox.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/child.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/clipped_multi_line_text_renderer.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/collapsing_header.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/combo.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/group.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/image.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/input_text.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/item_tooltip.cpp
+#        ${XFRAMES_BASE_PATH}/src/widget/map_view.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/multi_slider.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/plot_line.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/plot_candlestick.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/separator.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/separator_text.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/slider.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/table.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/tabs.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/text.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/text_wrap.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/tree_node.cpp
+        ${XFRAMES_BASE_PATH}/src/widget/window.cpp
 
 )
 
-add_library(reactDearImgui
-        ${REACT_IMGUI_SRC}
+add_library(xframes
+        ${XFRAMES_SRC}
         ${OSM_STATIC_MAP_GENERATOR_SRC}
         ${LEPTONICA_SRC}
         ${YOGA_SRC}
@@ -85,8 +85,8 @@ add_library(reactDearImgui
 )
 
 target_include_directories(
-        reactDearImgui PRIVATE
-        ${REACT_IMGUI_BASE_PATH}/include
+        xframes PRIVATE
+        ${XFRAMES_BASE_PATH}/include
         ${DEPS}/IconFontCppHeaders
         ${DEPS}/ReactivePlusPlus/src/rpp
 
@@ -98,4 +98,14 @@ target_include_directories(
         ${DEPS}/osm-static-map-generator/cpp/deps/leptonica/src
 )
 
-target_link_libraries(reactDearImgui PRIVATE imgui::imgui implot::implot nlohmann_json::nlohmann_json ada::ada JPEG::JPEG PNG::PNG TIFF::TIFF  ${CMAKE_CURRENT_SOURCE_DIR}/../glfw3dll.lib ${CMAKE_CURRENT_SOURCE_DIR}/../opengl32.lib)
+target_link_libraries(xframes PRIVATE
+        imgui::imgui
+        implot::implot
+        nlohmann_json::nlohmann_json
+        ada::ada
+        JPEG::JPEG
+        PNG::PNG
+        TIFF::TIFF
+        ${CMAKE_CURRENT_SOURCE_DIR}/../glfw3dll.lib
+        ${CMAKE_CURRENT_SOURCE_DIR}/../opengl32.lib
+)
