+++ packages/dear-imgui/cpp/node/.gitignore
@@ -1,6 +1,8 @@
 .idea
 build
 
+prebuilds
+
 imgui.ini
 
 glfw3.dll
