--- packages/dear-imgui/cpp/node/node-imgui.js
@@ -1,1 +1,0 @@
-module.exports = require("./build/Release/node-imgui.node")
