+++ packages/dear-imgui/cpp/node/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "node-imgui",
+  "name": "@xframes/node",
   "version": "0.0.1",
   "description": "A Node.js ImGui implementation",
   "main": "hello.js",
