+++ packages/dear-imgui/cpp/node/xframes.js
@@ -1,0 +1,1 @@
+module.exports = require("./build/Release/xframes.node")
