--- packages/dear-imgui/ts/node-imgui/tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES2022",
-        "module": "Node16",
-        "lib": ["ES2022"],
-        "moduleResolution": "Node16",
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "skipDefaultLibCheck": true,
-        "strict": true,
-        "baseUrl": "./",
-        "outDir": "./",
-        "forceConsistentCasingInFileNames": true,
-        "jsx": "react-jsx",
-        "allowJs": true
-    },
-    "include": ["./", "../src"],
-    "exclude": ["node_modules"]
-}
