+++ packages/dear-imgui/ts/package.json
@@ -1,11 +1,11 @@
 {
-    "name": "@react-wasm/dear-imgui",
-    "version": "0.3.5",
+    "name": "@xframes/wasm",
+    "version": "0.0.1",
     "main": "dist/index.js",
     "files": [
         "dist/*"
     ],
-    "homepage": "https://github.com/andreamancuso/react-wasm#readme",
+    "homepage": "https://github.com/andreamancuso/xframes#readme",
     "license": "MIT",
     "author": {
         "name": "Andrea Mancuso",
@@ -21,7 +21,7 @@
         "tsup": "tsup --format cjs --external react,react-dom",
         "copy-wasm-to-build-folder": "cpy --flat ./src/lib/wasm/reactDearImgui.* dist/",
         "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-        "node-imgui": "tsx ./node-imgui/index.tsx"
+        "node-imgui": "tsx ./xframes-node/index.tsx"
     },
     "tsup": {
         "entry": [
