+++ packages/dear-imgui/ts/xframes-node/TradingGuiDemo/cryptoSymbols.ts
@@ -1,0 +1,61 @@
+export const cryptoSymbols = [
+    "SUSHI/USD",
+    "YFI/USD",
+    "LTC/USD",
+    "DOGE/USD",
+    "BCH/USDT",
+    "BAT/USDC",
+    "BAT/USD",
+    "CRV/USDC",
+    "SHIB/USDT",
+    "BCH/USDC",
+    "LINK/USDT",
+    "LINK/USDC",
+    "BCH/USD",
+    "CRV/USD",
+    "AVAX/USDT",
+    "LTC/USDT",
+    "XTZ/USDC",
+    "BTC/USDT",
+    "AVAX/USDC",
+    "XTZ/USD",
+    // "AVAX/USD",
+    // "LTC/USDC",
+    // "LINK/USD",
+    // "LINK/BTC",
+    // "SHIB/USDC",
+    // "SOL/USDC",
+    // "GRT/USDC",
+    // "USDT/USDC",
+    // "SOL/USD",
+    // "SOL/USDT",
+    // "SHIB/USD",
+    // "DOGE/USDC",
+    // "GRT/USD",
+    // "ETH/USDT",
+    // "USDT/USD",
+    // "ETH/USDC",
+    // "LTC/BTC",
+    // "AAVE/USDT",
+    // "ETH/USD",
+    // "UNI/USDT",
+    // "ETH/BTC",
+    // "BCH/BTC",
+    // "USDC/USD",
+    // "YFI/USDT",
+    // "UNI/USDC",
+    // "DOT/USDC",
+    // "DOT/USD",
+    // "AAVE/USD",
+    // "MKR/USDC",
+    // "UNI/USD",
+    // "AAVE/USDC",
+    // "UNI/BTC",
+    // "SUSHI/USDT",
+    // "BTC/USDC",
+    // "BTC/USD",
+    // "SUSHI/USDC",
+    // "YFI/USDC",
+    // "MKR/USD",
+    // "DOGE/USDT",
+];
