+++ packages/dear-imgui/ts/xframes-node/TradingGuiDemo/dataServiceContext.tsx
@@ -1,0 +1,14 @@
+import { createContext, useContext } from "react";
+import { DataService } from "./dataService";
+
+export const DataServiceContext = createContext<DataService | null>(null);
+
+export const useDataService = () => {
+    const context = useContext(DataServiceContext);
+
+    if (context) {
+        return context;
+    }
+
+    throw new Error("DataServiceContext initialisation failed");
+};
