+++ packages/dear-imgui/ts/xframes-node/tsconfig.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+        "target": "ES2022",
+        "module": "Node16",
+        "lib": ["ES2022"],
+        "moduleResolution": "Node16",
+        "esModuleInterop": true,
+        "skipLibCheck": true,
+        "skipDefaultLibCheck": true,
+        "strict": true,
+        "baseUrl": "./",
+        "outDir": "./",
+        "forceConsistentCasingInFileNames": true,
+        "jsx": "react-jsx",
+        "allowJs": true
+    },
+    "include": ["./", "../src"],
+    "exclude": ["node_modules"]
+}
